Update ciudadFeriadoControlador.ts
Implementación de auditoria en los métodos del controlador
Implementación de transacciones

+++ servidor/src/controlador/ciudadFeriado/ciudadFeriadoControlador.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express';
 import { QueryResult } from 'pg';
+import AUDITORIA_CONTROLADOR from '../auditoria/auditoriaControlador';
 import pool from '../../database';
 
 class CiudadFeriadoControlador {
@@ -42,14 +43,61 @@ class CiudadFeriadoControlador {
     }
 
     // METODO PARA ELIMINAR REGISTRO
-    public async EliminarCiudadFeriado(req: Request, res: Response): Promise<void> {
-        const id = req.params.id;
-        await pool.query(
-            `
-            DELETE FROM ciud_feriados WHERE id = $1
-            `
-            , [id]);
-        res.jsonp({ message: 'Registro eliminado.' });
+    public async EliminarCiudadFeriado(req: Request, res: Response): Promise<Response> {
+        try {
+            // TODO ANALIZAR COMOOBTENER USER_NAME E IP DESDE EL FRONT
+            const { user_name, ip } = req.body;
+            const id = req.params.id;
+
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
+
+            // CONSULTAR DATOS ORIGINALES
+            const ciudad = await pool.query('SELECT * FROM ciud_feriados WHERE id = $1', [id]);
+            const [datosOriginales] = ciudad.rows;
+
+            if (!datosOriginales) {
+                // AUDITORIA
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'ciud_feriados',
+                    usuario: user_name,
+                    accion: 'D',
+                    datosOriginales: '',
+                    datosNuevos: '',
+                    ip,
+                    observacion: `Error al eliminar la ciudad con id ${id}`
+                });
+
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+                return res.status(404).jsonp({ message: 'Error al eliminar el registro.' });
+            }
+
+            await pool.query(
+                `
+                DELETE FROM ciud_feriados WHERE id = $1
+                `
+                , [id]);
+
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'ciud_feriados',
+                usuario: user_name,
+                accion: 'D',
+                datosOriginales: JSON.stringify(datosOriginales),
+                datosNuevos: '',
+                ip,
+                observacion: null
+            });
+
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
+            return res.jsonp({ message: 'Registro eliminado.' });
+        } catch (error) {
+            // REVERTIR TRANSACCION
+            await pool.query('ROLLBACK');
+            return res.status(404).jsonp({ message: 'Error al eliminar el registro.' });
+        }
     }
 
     // METODO PARA BUSCAR ID DE CIUDADES
@@ -71,7 +119,11 @@ class CiudadFeriadoControlador {
     // METODO PARA ASIGNAR CIUDADES A FERIADO
     public async AsignarCiudadFeriado(req: Request, res: Response) {
         try {
-            const { id_feriado, id_ciudad } = req.body;
+            const { id_feriado, id_ciudad, user_name, ip } = req.body;
+
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
+
             const response: QueryResult = await pool.query(
                 `
                 INSERT INTO ciud_feriados (id_feriado, id_ciudad) VALUES ($1, $2) RETURNING *
@@ -79,6 +131,20 @@ class CiudadFeriadoControlador {
                 , [id_feriado, id_ciudad]);
     
             const [feriado] = response.rows;
+
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'ciud_feriados',
+                usuario: user_name,
+                accion: 'I',
+                datosOriginales: '',
+                datosNuevos: `{id_feriado: ${id_feriado}, id_ciudad: ${id_ciudad}}`,
+                ip,
+                observacion: null
+            });
+
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
     
             if (feriado) {
                 return res.status(200).jsonp({ message: 'OK', reloj: feriado })
@@ -87,35 +153,68 @@ class CiudadFeriadoControlador {
                 return res.status(404).jsonp({ message: 'error' })
             }
         } catch (error) {
+            // REVERTIR TRANSACCION
+            await pool.query('ROLLBACK');
             return res.status(500).jsonp({ message: 'error' })
         }
 
     }
 
     // METODO PARA ACTUALIZAR REGISTRO
-    public async ActualizarCiudadFeriado(req: Request, res: Response): Promise<void> {
-        const { id_feriado, id_ciudad, id } = req.body;
-        await pool.query(
-            `
-            UPDATE ciud_feriados SET id_feriado = $1, id_ciudad = $2 WHERE id = $3
-            `
-            , [id_feriado, id_ciudad, id]);
-        res.jsonp({ message: 'Registro actualizado.' });
+    public async ActualizarCiudadFeriado(req: Request, res: Response): Promise<Response> {
+        try {
+            const { id_feriado, id_ciudad, id, user_name, ip } = req.body;
+    
+            // INICIAR TRANSACCION
+            await pool.query('BEGIN');
+    
+            // CONSULTAR DATOS ORIGINALES
+            const ciudad = await pool.query('SELECT * FROM ciud_feriados WHERE id = $1', [id]);
+            const [datosOriginales] = ciudad.rows;
+    
+            if (!datosOriginales) {
+                // AUDITORIA
+                await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                    tabla: 'ciud_feriados',
+                    usuario: user_name,
+                    accion: 'U',
+                    datosOriginales: '',
+                    datosNuevos: '',
+                    ip,
+                    observacion: `Error al actualizar la ciudad con id ${id}`
+                });
+    
+                // FINALIZAR TRANSACCION
+                await pool.query('COMMIT');
+                return res.status(404).jsonp({ message: 'Error al actualizar el registro.' });
+            }
+    
+            await pool.query(
+                `
+                UPDATE ciud_feriados SET id_feriado = $1, id_ciudad = $2 WHERE id = $3
+                `
+                , [id_feriado, id_ciudad, id]);
+    
+            // AUDITORIA
+            await AUDITORIA_CONTROLADOR.InsertarAuditoria({
+                tabla: 'ciud_feriados',
+                usuario: user_name,
+                accion: 'U',
+                datosOriginales: JSON.stringify(datosOriginales),
+                datosNuevos: `{id_feriado: ${id_feriado}, id_ciudad: ${id_ciudad}}`,
+                ip,
+                observacion: null
+            });
+    
+            // FINALIZAR TRANSACCION
+            await pool.query('COMMIT');
+            return res.jsonp({ message: 'Registro actualizado.' });
+        } catch (error) {
+            // FINALIZAR TRANSACCION
+            await pool.query('ROLLBACK');
+            return res.status(404).jsonp({ message: 'Error al actualizar el registro.' });
+        }
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 
     public async ObtenerFeriadosCiudad(req: Request, res: Response): Promise<any> {
         const id_ciudad = req.params.id_ciudad;
@@ -127,8 +226,6 @@ class CiudadFeriadoControlador {
             return res.status(404).jsonp({ text: 'Registros no encontrados' });
         }
     }
-
-
 
 }
